Updates for the new Release.  Set version number and release artifact Id.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/tags/activemq-cpp-2.2.2@719236 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/Message.cpp
@@ -458,10 +458,10 @@ bool Message::equals( const DataStructure* value ) const {
 ////////////////////////////////////////////////////////////////////////////////
 unsigned int Message::getSize() const {
 
-    long long size = DEFAULT_MESSAGE_SIZE;
+    unsigned int size = DEFAULT_MESSAGE_SIZE;
 
-    size += this->getContent().size();
-    size += this->getMarshalledProperties().size();
+    size += (unsigned int)this->getContent().size();
+    size += (unsigned int)this->getMarshalledProperties().size();
 
     return size;
 }
